+++ AuthPermissions.AspNetCore/OpenIdCode/AzureAdOpenIdExtension.cs
@@ -2,7 +2,10 @@
 // Licensed under MIT license. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
 using System.Security.Claims;
+using System.Threading.Tasks;
 using AuthPermissions.AdminCode;
 using AuthPermissions.CommonCode;
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
@@ -30,6 +33,7 @@ namespace AuthPermissions.AspNetCore.OpenIdCode
                     OnTokenValidated = async ctx => 
                     {
                         string userId = ctx.Principal.FindFirstValue(settings.UserIdClaimName);
+                        var email = ctx.Principal.FindFirstValue(settings.EmailClaimName);
 
                         var authPUserService =
                             ctx.HttpContext.RequestServices.GetRequiredService<IAuthUsersAdminService>();
@@ -42,7 +46,6 @@ namespace AuthPermissions.AspNetCore.OpenIdCode
 
                             if (settings.AddNewUserIfNotPresent)
                             {
-                                var email = ctx.Principal.FindFirstValue(settings.EmailClaimName);
                                 var username = ctx.Principal.FindFirstValue(settings.UsernameClaimName);
 
                                 var createStatus =
@@ -56,21 +59,43 @@ namespace AuthPermissions.AspNetCore.OpenIdCode
                                 logger.LogWarning($"A user with UserId = {userId} logged in, but was not in the AuthP user database.");
                             }
 
-                            return;
+                            //We replace some of the claims in the ClaimPrincipal so that the claims match what AuthP expects
+                            CreateClaimPrincipalWithAuthPClaims(ctx, userId, email);
                         }
+                        else
+                        {
+                            //We have an existing AuthP user, so we add their claims
+                            var claimsCalculator =
+                                ctx.HttpContext.RequestServices.GetRequiredService<IClaimsCalculator>();
 
-                        //We have an existing AuthP user, so we add their claims
-                        var claimsCalculator =
-                            ctx.HttpContext.RequestServices.GetRequiredService<IClaimsCalculator>();
-
-                        var claimsToAdd = await claimsCalculator.GetClaimsForAuthUserAsync(userId);
-                        var appIdentity = new ClaimsIdentity(claimsToAdd);
-                        ctx.Principal.AddIdentity(appIdentity);
-
-                        return;
+                            CreateClaimPrincipalWithAuthPClaims(ctx, userId, email, await claimsCalculator.GetClaimsForAuthUserAsync(userId));
+                        }
                     }
                 };
             });
+        }
+
+        private static void CreateClaimPrincipalWithAuthPClaims(TokenValidatedContext ctx,
+            string userId, string email, List<Claim> claimsToAdd = null)
+        {
+            var updatedClaims = ctx.Principal.Claims.ToList();
+            
+            if(claimsToAdd != null)
+                //add the AuthP claims
+                updatedClaims.AddRange(claimsToAdd);
+
+            //NOTE: The ClaimTypes.NameIdentifier is expected to contain the UserId, but with AzureId you get another value
+            //Therefore we remove/replace the NameIdentifier claim to have the user's id
+            updatedClaims.Remove(
+                updatedClaims.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier));
+            updatedClaims.Add(new Claim(ClaimTypes.NameIdentifier, userId));
+
+            //NOTE: We need to provide the Name claim to get the correct name shown in the ASP.NET Core sign in/sign out display
+            updatedClaims.Add(new Claim(ClaimTypes.Name, email));
+
+            //now we create a new ClaimsIdentity to replace the existing Principal
+            var appIdentity = new ClaimsIdentity(updatedClaims, ctx.Principal.Identity.AuthenticationType);
+            ctx.Principal = new ClaimsPrincipal(appIdentity);
         }
     }
 }