+++ Example5.MvcWebApp.AzureAdB2C/Controllers/HomeController.cs
@@ -2,16 +2,10 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
-using AuthPermissions.AdminCode;
 
 namespace Example5.MvcWebApp.AzureAdB2C.Controllers
 {
-    [Authorize]
     public class HomeController : Controller
     {
         private readonly ILogger<HomeController> _logger;
@@ -23,7 +17,7 @@ namespace Example5.MvcWebApp.AzureAdB2C.Controllers
 
         public IActionResult Index()
         {
-            return View();
+            return View(new AppSummary());
         }
 
         public IActionResult Privacy()