+++ Example5.MvcWebApp.AzureAdB2C/Startup.cs
@@ -47,7 +47,8 @@ namespace Example5.MvcWebApp.AzureAdB2C
             services.RegisterAuthPermissions<Example5Permissions>()
                 //NOTE: This uses the same database as the individual accounts DB
                 .UsingEfCoreSqlServer(Configuration.GetConnectionString("DefaultConnection"))
-                .AzureAdAuthentication(AzureAdSettings.AzureAdDefaultSettings())
+                //This uses the 
+                .AzureAdAuthentication(AzureAdSettings.AzureAdDefaultSettings(false))
                 .RegisterAuthenticationProviderReader<SyncAzureAdUsers>()
                 .AddRolesPermissionsIfEmpty(Example5AppAuthSetupData.BulkLoadRolesWithPermissions)
                 .AddAuthUsersIfEmpty(Example5AppAuthSetupData.UsersRolesDefinition)